+++ src/main/java/duke/command/ClearCommand.java
@@ -1,3 +1,9 @@
+package duke.command;
+
+import duke.TaskList;
+import duke.Ui;
+import duke.Storage;
+
 public class ClearCommand extends Command {
     public void execute(TaskList tasks, Ui ui, Storage storage) {
         tasks.clear();
