+++ src/main/java/duke/command/DeleteCommand.java
@@ -1,7 +1,14 @@
+package duke.command;
+
+import duke.Storage;
+import duke.TaskList;
+import duke.Ui;
+import duke.task.Task;
+
 public class DeleteCommand extends Command {
     private int taskNo;
 
-    DeleteCommand(int taskNo) {
+    public DeleteCommand(int taskNo) {
         this.taskNo = taskNo;
     }
