+++ src/main/java/duke/command/TodoCommand.java
@@ -1,16 +1,21 @@
-public class EventCommand extends Command {
-    private String title;
-    private String to;
-    private String from;
+package duke.command;
 
-    EventCommand(String title, String to, String from) {
+import duke.DukeException;
+import duke.Storage;
+import duke.TaskList;
+import duke.Ui;
+import duke.task.Task;
+import duke.task.Todo;
+
+public class TodoCommand extends Command {
+    private String title;
+
+    public TodoCommand(String title) {
         this.title = title;
-        this.to = to;
-        this.from = from;
     }
 
     public void execute(TaskList tasks, Ui ui, Storage storage) throws DukeException {
-        Task newTask = new Event(this.title, this.to, this.from);
+        Task newTask = new Todo(this.title);
         tasks.add(newTask);
         ui.showAdd(newTask);
         storage.save(tasks);
