+++ src/main/java/duke/enums/Commands.java
@@ -1,3 +1,5 @@
+package duke.enums;
+
 public enum Commands {
     TODO("todo"),
     DEADLINE("deadline"),
