+++ src/main/java/duke/enums/Views.java
@@ -1,3 +1,5 @@
+package duke.enums;
+
 public enum Views {
     LINE_STRING("____________________________________________________________"),
     WELCOME_STRING("Hello! I'm Duke\n      What can I do for grades?"),
